+++ test/schemas/test_body.py
@@ -11,5 +11,5 @@ def test_(case):
         """,
         **as_param({"name": "id", "in": "body", "required": True, "schema": {"type": "integer"}})
     )
-    # Then the generated test case should contain it its `body` attribute
+    # Then the generated test case should contain it in its `body` attribute
     testdir.run_and_assert("-s", passed=1)