+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Added
 ~~~~~
 
 - Open API 3.0 support. `#10`_
+- "header" parameters. `#7`_
 
 Changed
 ~~~~~~~
@@ -33,3 +34,4 @@ Fixed
 .. _#18: https://github.com/kiwicom/schemathesis/issues/18
 .. _#16: https://github.com/kiwicom/schemathesis/issues/16
 .. _#10: https://github.com/kiwicom/schemathesis/issues/10
+.. _#7: https://github.com/kiwicom/schemathesis/issues/7